AIO: Initial Implementation
This is an initial implementation of the ADC API for zmraa.  It works with the
nanokernel on the ARC core and has also been tested with the qmsi_ss driver on
the ARC core. These tests have been done on the Arduino 101 board.  It does not
currently work with the x86 core as it does not support analog.

Signed-off-by: Abhishek Malik <abhishek.malik@intel.com>
Signed-off-by: Brendan Le Foll <brendan.le.foll@intel.com>