PWM: Initial Implementation
This is the initial implementation of the pwm mraa api to be used with zephyr.
This is currently not fully tested however, most of the functionalities work as
expected.  Testing is in progress on the arduino 101. This implementation uses
the QMSI driver on both the arc and the quark cores. There are ifdefs for the
DW driver as well but it looks like the pwm dw driver is no longer supported by
zephyr. If required they can be removed with the next update.

Signed-off-by: Abhishek Malik <abhishek.malik@intel.com>
Signed-off-by: Brendan Le Foll <brendan.le.foll@intel.com>