SPI: Initial implementation
This commit adds the basic support for SPI in ZMRAA.
The driver was tested using the apa102 driver and also
a Logic Analyzer. This is just an initial implementation
and should work with v1.4 and v1.5. Not all the functions
have been tested so far, however, most of the functions
use the transceive function from the underlying API and
that function has been thoroughly tested.

Signed-off-by: Abhishek Malik <abhishek.malik@intel.com>