SPI: Initial Implementation
This is an initial SPI implementation. There is no direct
CS support on the 101 through SPI, however, it can be achieved by
using certain kernel config flags. An example is under works as of
now for SPI which will explain things further. This zmraa
driver has been tested with the apa102 C driver from UPM.

Signed-off-by: Abhishek Malik <abhishek.malik@intel.com>
Signed-off-by: Brendan Le Foll <brendan.le.foll@intel.com>