SPI: Enable SPI support on the ARC
This commits enables the SPI support on the ARC. Pinmuxing is not
required for SPI on the ARC because this is done by default by
Zephyr. SPI from the ARC is not exposed on the external headers,
but is useful as it is connected to the internal BMI160 on the
Arduino 101.

Signed-off-by: Abhishek Malik <abhishek.malik@intel.com>