D2000: Enable debugging via JTAG
Previously, the first thing mraa_init did was to mux TDO
to GPIO/PWM which terminates a JTAG debug session.  Updated
the board file to make sure JTAG (IO FUNCTION A) is enabled
for a DEBUG build.  Also, enable IO6 as a PWM ONLY IF CONFIG_PWM
has been set and not in DEBUG mode.  This makes it easier to
flash the firmware since a bootstall is required once the JTAG
pins have been muxed away.

Signed-off-by: Noel Eck <noel.eck@intel.com>